+++ tests/UnderstudyTest.php
@@ -8,6 +8,8 @@ use Rasuvaeff\Understudy\Arg;
 use Rasuvaeff\Understudy\Codegen\Blueprint;
 use Rasuvaeff\Understudy\Codegen\DoubleFactory;
 use Rasuvaeff\Understudy\Codegen\MethodSignature;
+use Rasuvaeff\Understudy\Codegen\TargetUnifier;
+use Rasuvaeff\Understudy\Codegen\TypeRenderer;
 use Rasuvaeff\Understudy\Exception\ForgottenDouble;
 use Rasuvaeff\Understudy\Exception\InvalidCallSpecification;
 use Rasuvaeff\Understudy\Exception\MatcherLeaked;
@@ -75,6 +77,8 @@ use function Rasuvaeff\Understudy\when;
 #[Covers(StrictModeViolation::class)]
 #[Covers(UnsupportedTarget::class)]
 #[Covers(VerificationFailed::class)]
+#[Covers(TargetUnifier::class)]
+#[Covers(TypeRenderer::class)]
 final class UnderstudyTest
 {
     #[AfterTest]